backend/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
python-dotenv==0.19.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
redis==4.3.4
httpx==0.25.1
requests==2.31.0
PyJWT==2.8.0
pinecone-client==2.2.4 